\name{evirBuiltin}

\alias{evir}

\alias{decluster} 
\alias{emplot}
\alias{exindex}
\alias{findthresh} 
\alias{gev} 
\alias{gev.dens} 
\alias{gev.diag} 
\alias{gev.fit}
\alias{gev.his}
\alias{gev.pp} 
\alias{gev.prof} 
\alias{gev.profxi} 
\alias{gev.qq} 
\alias{gev.rl} 
\alias{gevf} 
\alias{gevq} 
\alias{gpd} 
\alias{gpd.dens} 
\alias{gpd.diag} 
\alias{gpd.fit} 
\alias{gpd.fitrange} 
\alias{gpd.his} 
\alias{pot}
\alias{gpd.pp} 
\alias{gpd.prof} 
\alias{gpd.profxi}
\alias{gpd.q} 
\alias{gpd.qq} 
\alias{gpd.rl} 
\alias{gpd.sfall} 
\alias{gpdbiv} 
\alias{gpdf} 
\alias{gpdq} 
\alias{gpdq2} 
\alias{gum.dens} 
\alias{gum.df}
\alias{gum.diag} 
\alias{gum.fit} 
\alias{gum.q} 
\alias{gum.rl} 
\alias{gumbel} 
\alias{hill} 
\alias{identity} 
\alias{interpret.gpdbiv}
\alias{meplot} 
\alias{mrl.plot} 
\alias{plot.gev} 
\alias{plot.gpd} 
\alias{plot.gpdbiv} 
\alias{plot.pot} 
\alias{pp.diag}
\alias{pp.fit} 
\alias{pp.fitrange} 
\alias{pp.pp} 
\alias{pp.qq} 
\alias{ppf}
\alias{ppp} 
\alias{ppq} 
\alias{q.form} 
\alias{qplot} 
\alias{quant}
\alias{records} 
\alias{riskmeasures} 
\alias{rlarg.diag} 
\alias{rlarg.fit} 
\alias{rlarg.pp} 
\alias{rlarg.qq} 
\alias{rlargf}
\alias{rlargq} 
\alias{rlargq2} 
\alias{rlevel.gev} 
\alias{shape} 
\alias{tailplot}


\alias{bmw}
\alias{danish} 
\alias{dowjones} 
\alias{engine} 
\alias{euroex} 
\alias{exchange} 
\alias{fremantle} 
\alias{glass}
\alias{nidd.annual} 
\alias{nidd.thresh} 
\alias{portpirie}
\alias{rain} 
\alias{siemens} 
\alias{sp.raw} 
\alias{spto87} 
\alias{venice}
\alias{wavesurge} 
\alias{wind} 
\alias{wooster}


\alias{bmw.ret}
\alias{dax.ret}


\title{evir Builtin Functions}


\description{

    Builtin functions as available in the contributed 
    \R-packages 'evir' and 'ismev'. They are used only 
    as internal functions, and they are not thought 
    for usage by Rmetrics users.
    
}


\examples{
## SOURCE("fExtremes.056A-ExtremesBuiltin")

## -
}


\keyword{data}
