\name{fExtremesData}

\alias{fExtremesData}

\alias{bmwRet}
\alias{danishClaims}
\alias{daxRet}
\alias{bmw}
\alias{danish}
\alias{dowjones}
\alias{engine}
\alias{euroex}
\alias{exchange}
\alias{fremantle}
\alias{glass}
\alias{nidd.annual}
\alias{nidd.thresh}
\alias{portpirie}
\alias{rain}
\alias{siemens}
\alias{sp.raw}
\alias{spto87}
\alias{venice}
\alias{wavesurge}
\alias{wind}
\alias{wooster}


\title{fExtremes Data Sets}


\description{
        
    A collection and description of public data sets 
    used by fExtremes and furter data sets form the 
    contributed R packages evir and ismev.
    \cr

}


\details{

    For details we rever to the help pages of the
    contributed R packages evir and ismev.
        
}


\keyword{datasets}

